feat(gwasUnifiedFlowFileOrganization): Renamed CohortSelect to SelectCohort to match 'SelectOutcome'
