+++ src/Analysis/GWASV2/GWASContainer.jsx
@@ -18,47 +18,47 @@ const GWASContainer = () => {
 
   const generateStep = () => {
     switch (state.currentStep) {
-      case 0:
-        return (
-          <SelectStudyPopulation
-            selectedCohort={state.selectedStudyPopulationCohort}
-            dispatch={dispatch}
-          />
-        );
-      case 1:
-        return (
-          <SelectOutcome
-            outcome={state.outcome}
+    case 0:
+      return (
+        <SelectStudyPopulation
+          selectedCohort={state.selectedStudyPopulationCohort}
+          dispatch={dispatch}
+        />
+      );
+    case 1:
+      return (
+        <SelectOutcome
+          outcome={state.outcome}
+          covariates={state.covariates}
+          dispatch={dispatch}
+        />
+      );
+    case 2:
+      return (
+        <React.Fragment>
+          <SelectCovariates
+            outcome={{}}
             covariates={state.covariates}
             dispatch={dispatch}
           />
-        );
-      case 2:
-        return (
-          <React.Fragment>
-            <SelectCovariates
-              outcome={{}}
-              covariates={state.covariates}
-              dispatch={dispatch}
-            />
-            <CovariatesCardsList
-              covariates={state.covariates}
-              dispatch={dispatch}
-            />
-          </React.Fragment>
-        );
-      case 3:
-        return (
-          <ConfigureGWAS
+          <CovariatesCardsList
+            covariates={state.covariates}
             dispatch={dispatch}
-            numOfPCs={state.numPCs}
-            mafThreshold={state.mafThreshold}
-            imputationScore={state.imputationScore}
-            selectedHare={state.selectedHare}
           />
-        );
-      default:
-        return null;
+        </React.Fragment>
+      );
+    case 3:
+      return (
+        <ConfigureGWAS
+          dispatch={dispatch}
+          numOfPCs={state.numPCs}
+          mafThreshold={state.mafThreshold}
+          imputationScore={state.imputationScore}
+          selectedHare={state.selectedHare}
+        />
+      );
+    default:
+      return null;
     }
   };
