+++ src/Analysis/GWASV2/Steps/SelectStudyPopulation/SelectStudyPopulation.jsx
@@ -1,7 +1,7 @@
 import React from 'react';
 import PropTypes from 'prop-types';
 import ACTIONS from '../../Shared/StateManagement/Actions';
-import CohortSelect from '../../CohortSelect/CohortSelect';
+import CohortSelect from '../../SelectCohort/SelectCohort';
 
 const SelectStudyPopulation = ({ selectedCohort, dispatch }) => {
   const handleStudyPopulationSelect = (selectedRow) => {
